+++ intermediate-language/src/main/scala/com/ing/baker/il/failurestrategy/BlockInteraction.scala
@@ -1,5 +1,0 @@
-package com.ing.baker.il.failurestrategy
-
-case object BlockInteraction extends InteractionFailureStrategy {
-  def apply(n: Int) : ExceptionStrategyOutcome = ExceptionStrategyOutcome.BlockTransition
-}