+++ intermediate-language/src/main/scala/com/ing/baker/il/failurestrategy/InteractionFailureStrategy.scala
@@ -1,5 +1,0 @@
-package com.ing.baker.il.failurestrategy
-
-trait InteractionFailureStrategy {
-  def apply(n: Int): ExceptionStrategyOutcome
-}